requirements.txt
phidata
yfinance
python-dotenv
fpdf
DuckDuckGo
logging
phi